+++ src/string/define.js
@@ -16,6 +16,7 @@ defineInstance('isEmpty', methods.isEmpty);
 defineInstance('last', methods.last);
 defineInstance('pad', methods.pad);
 defineInstance('parameterize', methods.parameterize);
+defineInstance('remove', methods.remove);
 defineInstance('reverse', methods.reverse);
 defineInstance('spacify', methods.spacify);
 defineInstance('stripTags', methods.stripTags);